Refactor and use require.js
This commit breaks up a single monolithic js file into 6 parts,
according to functionality. It also fixes bugs found in the course of
refactoring.
