+++ .eslintrc.js
@@ -20,6 +20,7 @@ module.exports = {
         "gtag": true,
         "onYouTubeIframeAPIReady": true,
         "YT": true,
+        "requirejs": true,
     },
     "rules": {
         "indent": [
