+++ themes/ohcoe/static/js/src/scores.js
@@ -1,0 +1,47 @@
+/* eslint-disable scanjs-rules/identifier_localStorage */
+define(function(){
+    function render() {
+        /* Display Results to Users */
+        $('.domain-results').each(function(idx, domainResults){
+            var resultsContainer = $(domainResults).find('tbody').first();
+            var currentDomain = domainResults.id;
+
+            [...Array(3).keys()].map((el) => {
+                var q = el + 1;
+                var prefix = currentDomain + '-' + q;
+                var preScore = localStorage.getItem(prefix + '-pre-q1');
+                var postScore = localStorage.getItem(prefix + '-post-q1');
+
+                if (preScore && postScore) {
+                    preScore = Number(preScore);
+                    postScore = Number(postScore);
+
+                    var learningObjectiveScore = postScore - preScore;
+                    var tableRow = $('<tr></tr>');
+                    tableRow.html(
+                        /* eslint-disable-next-line max-len */
+                        '<td><a href="' + localStorage.getItem(prefix + '-url') + '">' +
+                        localStorage.getItem(prefix) + '</a>' +
+                        /* eslint-disable-next-line max-len */
+                        '<a href="' + localStorage.getItem(prefix + '-url') + '" class="btn btn-success btn-sm">' +
+                        'Review Learning Objective</a>' +
+                        '</td>' +
+                        '<td>' + preScore + '</td>' +
+                        '<td>' + postScore + '</td>' +
+                        '<td>' + learningObjectiveScore + '</td>');
+                    resultsContainer.append(tableRow[0]);
+
+                    // Unhide the parent contains if there's at least one
+                    // question answered.
+                    $('#' + (idx + 1)).show();
+                    $('#domain-score-' + (idx + 1)).show();
+                    $('#domain-title-' + (idx + 1)).show();
+                }
+            });
+        });
+
+    }
+    return {
+        render: render,
+    };
+});
