+++ themes/ohcoe/static/js/src/utils.js
@@ -1,0 +1,25 @@
+define(function(){
+    /* Utility Methods */
+    function round(val){
+        return Math.round(val * 100) / 100;
+    }
+
+    function mean(arr){
+        return arr.reduce(function(acc, val){
+            return acc + val;
+        }) / arr.length;
+    }
+
+    function sanitize(s) {
+        // http://shebang.brandonmintern.com/foolproof-html-escaping-in-javascript/
+        var div = document.createElement('div');
+        div.appendChild(document.createTextNode(s));
+        return div.innerHTML;
+    }
+
+    return {
+        round: round,
+        mean: mean,
+        sanitize: sanitize,
+    };
+});
